+++ test-app/config/environment.js
@@ -18,9 +18,6 @@ module.exports = function (environment) {
     },
 
     APP: {
-      primitiveUI: {
-        portalElement: '#portal',
-      },
       // Here you can pass flags/options to your application instance
       // when it is created
     },
@@ -44,7 +41,6 @@ module.exports = function (environment) {
 
     ENV.APP.rootElement = '#ember-testing';
     ENV.APP.autoboot = false;
-    ENV.APP.primitiveUI.portalElement = ENV.APP.rootElement;
   }
 
   if (environment === 'production') {