+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `data-*` attribute to `Menu` component (`data-state="open"`, `data-state="active"`, etc).
 - `Popover` component.
+- `Menu` component yields `id`, `buttonId`, `listId`, `itemId`, `close`, `open` properties.
 
 ## [0.2.0] - 2023-04-30