+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Popover` component.
 - `Menu` component yields `id`, `buttonId`, `listId`, `itemId`, `close`, `open` properties.
 
+### Removed
+- Boilerplate setup for `Portal` component.
+
 ## [0.2.0] - 2023-04-30
 
 ### Added