+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `data-*` attribute to `Menu` component (`data-state="open"`, `data-state="active"`, etc).
 - `Popover` component.
 - `Menu` component yields `id`, `buttonId`, `listId`, `itemId`, `close`, `open` properties.
+- Set `data-placement` on velcro element.
 
 ## [0.2.0] - 2023-04-30